refactor(listener): extract skill and agent commands
